requirements.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-xdist>=3.5.0
httpx>=0.27.0
anyio>=4.0.0
allure-pytest>=2.13.0
python-dotenv>=1.0.0
